refactor: move utils to their own files
